007-05 gemini_asst_memoonly/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
notion-client>=2.0.0
pydantic>=2.8.0
langchain>=0.2.3
langchain-google-genai>=1.0.7
google-generativeai>=0.5.4
